refactor: 기존의 enum 클래스를 하나의 folder에서 관리하기 위해 수정
