+++ src/main/java/webserver/HttpConstants/HttpStatus.java
@@ -1,4 +1,4 @@
-package webserver.response;
+package webserver.HttpConstants;
 
 public enum HttpStatus {
